Read Description
1. CRUD and Add 3 columns to Candidate Detail (experience_year, industry and cv_file)
2. View Bootcamp Profile in Admin [DONE]
3. Candidate Detail Form Dynamic 30%

+++ database/migrations/2021_09_19_155128_create_education_changes_table.php
@@ -22,8 +22,8 @@ class CreateEducationChangesTable extends Migration
             $table->string('degree')->nullable();
             $table->string('school')->nullable();
             $table->string('major')->nullable();
-            $table->bigInteger('start_year')->nullable();
-            $table->bigInteger('end_year')->nullable();
+            $table->unsignedInteger('start_year')->nullable();
+            $table->unsignedInteger('end_year')->nullable();
             $table->enum('action', ['create', 'update', 'delete']);
             $table->timestamps();
         });
